requirements.txt
streamlit
python-dotenv
requests
langgraph
faiss-cpu
sentence-transformers
beautifulsoup4
pypdf
numpy